Rename Constants back to DecimalConstants Removed unnecessary set on ExchangeSpecificParameters map
